{"schema":"decodex/commit/1","summary":"split cli support modules","authority":"manual"}
