+++ apps/radar/src/lib.rs
@@ -3,6 +3,7 @@
 mod artifact_validation;
 mod cli;
 mod config;
+mod constants;
 mod core_io;
 mod github_api;
 mod github_bundle_client;
@@ -22,6 +23,20 @@ mod prelude {
 }
 
 pub(crate) use self::{
+	constants::{
+		ANALYSIS_DRAFT_KIND, ARTIFACT_KINDS, ATTENTION_RULES, BUNDLE_SCHEMA,
+		CONFIG_FEATURE_CATALOG_PATH, CONFIG_FEATURE_CATALOG_SCHEMA,
+		CONTROL_PLANE_UPGRADE_CANDIDATE_SCHEMA, DEFAULT_LEDGER_PATH, DEFAULT_MIN_STABLE_TAG,
+		DEFAULT_PAIR_LIMIT, DEFAULT_PREVIEW_LIMIT, DEFAULT_QUEUE_OUT, DEFAULT_RELEASE_DELTA_OUT,
+		DEFAULT_SEARCH_LIMIT, DEFAULT_SIGNALS_DIR, DEFAULT_STABLE_LIMIT, DEFAULT_TAG_PREFIX,
+		DEFAULT_VALIDATION_PATHS, GENERIC_COMMIT_TITLES, GITHUB_REQUEST_ATTEMPTS,
+		GITHUB_REQUEST_BACKOFF, GITHUB_REQUEST_TIMEOUT, HIGH_VALUE_SURFACES,
+		RADAR_ARCHIVE_HISTORICAL_RETENTION_CUTOFF, RELEASE_DELTA_SCHEMA,
+		RETRYABLE_GITHUB_STATUS_CODES, REVIEW_STATUSES, RUN_CODEX_ANALYSIS_SCRIPT, SCHEMA_VERSION,
+		SIGNAL_CONFIDENCE, SIGNAL_SCHEMA, SURFACE_RULES, UPSTREAM_IMPACT_SCHEMA,
+		UPSTREAM_REVIEW_LINEAR_FOLLOWUP_CUTOFF, UPSTREAM_REVIEW_QUEUE_SCHEMA,
+		UPSTREAM_REVIEW_SCHEMA, UPSTREAM_SUBJECT_KINDS,
+	},
 	core_io::{
 		absolute_repo_path, collect_bundle_json_files, ledger_path, load_known_feature_names,
 		repo_default_branch, sorted_json_files, validate_expected_schema,
@@ -61,11 +76,9 @@ use std::{
 	iter,
 	path::{Path, PathBuf},
 	process,
-	time::Duration,
 };
 
 use clap::Parser as _;
-use reqwest::StatusCode;
 use serde_json::{self, Map, Value};
 use time::{OffsetDateTime, format_description::well_known::Rfc3339};
 
@@ -85,111 +98,6 @@ use review_queue::{RecentCommit, build_review_queue};
 use signal_render::{rendered_config_flags, rendered_signal};
 use source_bundle::{build_commit_bundle_from_sources, build_pr_bundle_from_sources};
 
-const BUNDLE_SCHEMA: &str = "github_change_bundle/v1";
-const CONTROL_PLANE_UPGRADE_CANDIDATE_SCHEMA: &str = "control_plane_upgrade_candidate/v1";
-const DEFAULT_LEDGER_PATH: &str = paths::DEFAULT_LEDGER_PATH;
-const DEFAULT_MIN_STABLE_TAG: &str = "rust-v0.116.0";
-const DEFAULT_PAIR_LIMIT: usize = 24;
-const DEFAULT_PREVIEW_LIMIT: usize = 0;
-const DEFAULT_QUEUE_OUT: &str = paths::DEFAULT_QUEUE_OUT;
-const DEFAULT_RELEASE_DELTA_OUT: &str = paths::DEFAULT_RELEASE_DELTA_OUT;
-const DEFAULT_SEARCH_LIMIT: usize = 40;
-const DEFAULT_SIGNALS_DIR: &str = paths::DEFAULT_SIGNALS_DIR;
-const DEFAULT_STABLE_LIMIT: usize = 0;
-const DEFAULT_TAG_PREFIX: &str = "rust-v";
-const RELEASE_DELTA_SCHEMA: &str = "release_delta/v1";
-const SCHEMA_VERSION: i64 = 5;
-const SIGNAL_SCHEMA: &str = "signal_entry/v1";
-const UPSTREAM_IMPACT_SCHEMA: &str = "upstream_impact/v1";
-const UPSTREAM_REVIEW_QUEUE_SCHEMA: &str = "upstream_review_queue/v1";
-const UPSTREAM_REVIEW_SCHEMA: &str = "upstream_review/v1";
-const CONFIG_FEATURE_CATALOG_SCHEMA: &str = "codex_config_feature_catalog/v1";
-const ANALYSIS_DRAFT_KIND: &str = "analysis_draft";
-const RADAR_ARCHIVE_HISTORICAL_RETENTION_CUTOFF: &str = "2026-06-07T00:00:00Z";
-const UPSTREAM_REVIEW_LINEAR_FOLLOWUP_CUTOFF: &str = "2026-06-12T00:00:00Z";
-const SIGNAL_CONFIDENCE: &[&str] = &["confirmed", "likely", "weak"];
-const UPSTREAM_SUBJECT_KINDS: &[&str] = &["commit", "pr"];
-const DEFAULT_VALIDATION_PATHS: &[&str] = paths::DEFAULT_VALIDATION_PATHS;
-const GENERIC_COMMIT_TITLES: &[&str] =
-	&["update", "fix", "fix.", "fix tests", "fix tests.", "merge fixes", "flaky syntax"];
-const CONFIG_FEATURE_CATALOG_PATH: &str = paths::CONFIG_FEATURE_CATALOG_PATH;
-const RUN_CODEX_ANALYSIS_SCRIPT: &str = paths::RUN_CODEX_ANALYSIS_SCRIPT;
-const HIGH_VALUE_SURFACES: &[&str] = &[
-	"app_server_protocol",
-	"mcp_plugins",
-	"browser_chrome",
-	"sandbox_permissions",
-	"config_hooks",
-	"auth_accounts",
-	"model_provider",
-];
-const ATTENTION_RULES: &[(&str, &[&str])] = &[
-	(
-		"new_feature",
-		&["feat", "feature", "add ", "adds ", "support", "enable", "implement", "introduce"],
-	),
-	("deprecated_removed", &["deprecat", "remove", "removed", "delete", "disable", "no longer"]),
-	(
-		"protocol_change",
-		&[
-			"protocol",
-			"schema",
-			"api",
-			"json-rpc",
-			"jsonrpc",
-			"notification",
-			"request",
-			"response",
-		],
-	),
-	("breaking_change", &["breaking", "break ", "rename", "migration", "incompat", "no longer"]),
-	(
-		"security_policy",
-		&["sandbox", "permission", "approval", "full access", "network", "denylist", "allowlist"],
-	),
-	("rate_limit", &["rate limit", "ratelimit", "quota", "usage limit", "message cap"]),
-	("auth_account", &["auth", "account", "login", "token"]),
-	("release_packaging", &["release", "appcast", "sparkle", "beta", "version"]),
-];
-const SURFACE_RULES: &[(&str, &[&str])] = &[
-	("app_server_protocol", &["app-server", "app_server", "protocol", "jsonrpc", "json-rpc"]),
-	("mcp_plugins", &["mcp", "plugin", "tool-search", "tool_search"]),
-	("browser_chrome", &["browser", "chrome", "webview"]),
-	(
-		"sandbox_permissions",
-		&["sandbox", "permission", "approval", "policy", "denylist", "allowlist"],
-	),
-	("config_hooks", &["config", "hook", "settings", "toml"]),
-	("auth_accounts", &["auth", "account", "login", "token"]),
-	("model_provider", &["model", "provider", "rate-limit", "ratelimit", "quota"]),
-	("cli_tui", &["cli", "tui", "terminal", "chatwidget"]),
-	("release_packaging", &["release", "appcast", "sparkle", "version", "install", "package"]),
-	("docs_examples", &["docs/", "readme", "example"]),
-	("tests_ci", &["test", "tests", ".github", "ci", "fixture"]),
-];
-const REVIEW_STATUSES: &[&str] =
-	&["archived", "control_plane", "deprecated", "seen", "signal", "skipped", "watch"];
-const ARTIFACT_KINDS: &[&str] = &[
-	"analysis",
-	"archive_manifest",
-	"bundle",
-	"control_plane_upgrade_candidate",
-	"ledger_export",
-	"release_delta",
-	"signal",
-	"upstream_impact",
-];
-const GITHUB_REQUEST_ATTEMPTS: usize = 4;
-const GITHUB_REQUEST_BACKOFF: Duration = Duration::from_secs(1);
-const GITHUB_REQUEST_TIMEOUT: Duration = Duration::from_secs(30);
-const RETRYABLE_GITHUB_STATUS_CODES: &[StatusCode] = &[
-	StatusCode::TOO_MANY_REQUESTS,
-	StatusCode::INTERNAL_SERVER_ERROR,
-	StatusCode::BAD_GATEWAY,
-	StatusCode::SERVICE_UNAVAILABLE,
-	StatusCode::GATEWAY_TIMEOUT,
-];
-
 #[derive(Debug)]
 enum RefreshKind {
 	Queue,
@@ -203,24 +111,5 @@ pub fn run() -> Result<()> {
 	Cli::parse().run()
 }
 
-#[cfg(test)]
-mod test_support {
-	use std::sync::{Mutex, MutexGuard, OnceLock};
-
-	pub(crate) struct TestEnvLockGuard {
-		_lock: MutexGuard<'static, ()>,
-	}
-
-	pub(crate) fn lock_test_env() -> TestEnvLockGuard {
-		TestEnvLockGuard {
-			_lock: test_env_mutex().lock().expect("test env mutex should not be poisoned"),
-		}
-	}
-
-	fn test_env_mutex() -> &'static Mutex<()> {
-		static TEST_ENV_MUTEX: OnceLock<Mutex<()>> = OnceLock::new();
-
-		TEST_ENV_MUTEX.get_or_init(|| Mutex::new(()))
-	}
-}
+#[cfg(test)] mod test_support;
 #[cfg(test)] mod tests;
